fix: api.js in typescript version (#40)
Hi,

I got this error:
```
The requested module 'multiformats' does not provide an export named 'Block' from pail
```
this was happening because, besides api.js, the .d.ts was not in the same directory, which confused tsx.

To fix it I changed the exports in package.json and the emitRule of tsc to only use the emitted file structure.

SideTrack:

I updated the package.json, and with that, the test infra to use vitest. I hope that is not a problem -- 
if you want to stick with mocha and coverage -> I am not happy but could rewind -> but than the 
typescript is required to the test infra -> due to the removal of the api.js which causes this problem in the
first place.

I run standard --fix, which also fixes the formatting that's why so much files.